+++ CHANGES.rst
@@ -7,8 +7,7 @@ in progress
 ===========
 - Re-add export/influx.py lost during refactoring
 - Fix InfluxDB UDP data acquisition
-
-
+- Tap into error signalling over MQTT
 
 
 .. _kotori-0.19.1: